Biomes implemented blocks: leaves, grass, vines and lilypads

+++ colors.h
@@ -9,7 +9,7 @@
 // Colors are stored twice since BMP and PNG need them in different order
 // Noise is supposed to look normal when -noise 10 is given
 extern uint8_t colors[65536][8];
-extern uint8_t biomes[256][4];
+extern int16_t biomes[256][4];
 extern uint8_t colorsToMap[65536];
 extern uint16_t colorsToID[256];
